+++ src/app/layout.tsx
@@ -1,4 +1,5 @@
 import { Metadata } from 'next';
+import { ReactNode } from 'react';
 
 import '@/styles/globals.css';
 
@@ -29,11 +30,7 @@ export const metadata: Metadata = {
   },
 };
 
-export default function RootLayout({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
+export default function RootLayout({ children }: { children: ReactNode }) {
   return (
     <html>
       <body>{children}</body>
