+++ src/components/HeaderAndMainClaim.tsx
@@ -4,10 +4,11 @@ import { CgMouse } from 'react-icons/cg';
 
 import { isMobile } from '@/lib/utils';
 
+import CyanBox from '@/components/CyanBox';
 import Header from '@/components/Header';
 import MainClaim from '@/components/MainClaim';
 
-export default function CyanBoxAndHeader() {
+export default function HeaderAndMainClaim() {
   const [headerOpacity, setHeaderOpacity] = useState(0);
   const [scrollDownButtonOpacity, setScrollDownButtonOpacity] = useState(1);
 
@@ -45,31 +46,31 @@ export default function CyanBoxAndHeader() {
   }
 
   return (
-    <div className='sticky -top-[540px] z-10 lg:-top-[690px]'>
-      <div className='flex w-full justify-center'>
-        <div className='bg-primary-cyan custom-angled-rectangle-cyan-animated-mobile lg:custom-angled-rectangle-cyan-animated flex h-[700px] w-full flex-col justify-center lg:h-[850px] lg:w-[1300px]'>
-          <Header className='absolute top-0' />
-          {headerOpacity > 0 && (
-            <Header opacity={headerOpacity} className='fixed top-0 lg:mt-5' />
-          )}
-          <div className='flex h-full w-full flex-col justify-center'>
-            <MainClaim />
-            {scrollDownButtonOpacity > 0 && (
-              <button
-                style={{ opacity: scrollDownButtonOpacity }}
-                className='absolute bottom-12 flex max-w-[1300px] flex-col items-center gap-1 pl-6 text-white lg:gap-4 lg:pl-24'
-                onClick={scrollDown}
-              >
-                <div className='text-xs font-light [writing-mode:vertical-lr]'>
-                  WHAT WE DO
-                </div>
-                <CgMouse className='hidden lg:flex' size={40} />
-                <BsArrowDownShort className='flex lg:hidden' size={40} />
-              </button>
-            )}
-          </div>
-        </div>
+    <CyanBox className='sticky'>
+      <div className='flex h-full w-full flex-col justify-center'>
+        <Header className='absolute top-0' clickOnLogoBehaviour='scrollUp' />
+        {headerOpacity > 0 && (
+          <Header
+            opacity={headerOpacity}
+            className='absolute bottom-10'
+            clickOnLogoBehaviour='scrollUp'
+          />
+        )}
+        <MainClaim />
+        {scrollDownButtonOpacity > 0 && (
+          <button
+            style={{ opacity: scrollDownButtonOpacity }}
+            className='absolute bottom-12 flex max-w-[1300px] flex-col items-center gap-1 pl-6 text-white lg:gap-4 lg:pl-24'
+            onClick={scrollDown}
+          >
+            <div className='text-xs font-light [writing-mode:vertical-lr]'>
+              WHAT WE DO
+            </div>
+            <CgMouse className='hidden lg:flex' size={40} />
+            <BsArrowDownShort className='flex lg:hidden' size={40} />
+          </button>
+        )}
       </div>
-    </div>
+    </CyanBox>
   );
 }
